+++ language/bert/basic/utils/tokenizer.py
@@ -1,0 +1,131 @@
+'''
+Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of this
+software and associated documentation files (the "Software"), to deal in the Software
+without restriction, including without limitation the rights to use, copy, modify,
+merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
+permit persons to whom the Software is furnished to do so.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+'''
+
+import random
+import torch
+
+# Get tokens of the BERT input sequence and their segment IDs
+def _get_tokens_and_segments(tokens_a, tokens_b = None):
+    tokens = ['<cls>'] + tokens_a + ['<sep>']
+    # 0 and 1 are marking segment A and B respectively
+    segments = [0] * (len(tokens_a) + 2)
+    if tokens_b is not None:
+        tokens += tokens_b + ['<sep>']
+        segments += [1] * (len(tokens_b) + 1)
+    return tokens, segments
+
+# Generating the Next Sentence Prediction(nsp) Task
+def _get_next_sentence(sentence, next_sentence, paragraphs):
+    if random.random() < 0.5:
+        is_next = True
+    else:
+        next_sentence = random.choice(random.choice(paragraphs))
+        is_next = False
+    return sentence, next_sentence, is_next
+
+def get_nsp_data_from_paragraph(paragraph, paragraphs, max_len):
+    nsp_data_from_paragraph = []
+    for i in range(len(paragraph) - 1):
+        tokens_a, tokens_b, is_next = _get_next_sentence(paragraph[i], paragraph[i+1], paragraphs)
+        # 1 '<cls>' token and 2 '<sep>' tokens
+        if len(tokens_a) + len(tokens_b) + 3 > max_len:
+            continue
+        tokens, segments = _get_tokens_and_segments(tokens_a, tokens_b)
+        nsp_data_from_paragraph.append((tokens, segments, is_next))
+    return nsp_data_from_paragraph
+
+# Generating the Masked Language Modeling(mlm) Task
+# tokens: a list of tokens representing a BERT input sequence
+# candidate_pred_positions: a list of token indices of the BERT input sequence excluding those of special tokens
+# num_mlm_preds: the number of predictions (recall 15% random tokens to predict)
+def _replace_mlm_tokens(tokens, candidata_pred_postions, num_mlm_preds, vocab):
+    # The input may contain replaced '<mask>' or random tokens
+    mlm_input_tokens = tokens[:]
+    pred_positions_and_labels = []
+    # Shuffle for getting 15% random tokens for prediction
+    random.shuffle(candidata_pred_postions)
+    for mlm_pred_postion in candidata_pred_postions:
+        if len(pred_positions_and_labels) >= num_mlm_preds:
+            break
+        masked_token = None
+        # 80% replace the word with the '<mask>' token
+        if random.random() < 0.8:
+            masked_token = '<mask>'
+        else:
+            # 10% keep the word unchanged
+            if random.random() < 0.5:
+                masked_token = tokens[mlm_pred_postion]
+            # 10% replace the word with a random word
+            else:
+                masked_token = random.choice(vocab.idx_to_token)
+        mlm_input_tokens[mlm_pred_postion] = masked_token
+        pred_positions_and_labels.append((mlm_pred_postion, tokens[mlm_pred_postion]))
+    return mlm_input_tokens, pred_positions_and_labels
+
+def get_mlm_data_from_tokens(tokens, vocab):
+    candidate_pred_postions = []
+    for i, token in enumerate(tokens):
+        # Ignore special tokens
+        if token in ['<cls>', '<sep>']:
+            continue
+        candidate_pred_postions.append(i)
+    # 15% of random tokens are predicted in the masked language modeling task
+    num_mlm_preds = max(1, round(len(tokens) * 0.15))
+    mlm_input_tokens, pred_positions_and_labels = _replace_mlm_tokens(tokens, 
+        candidate_pred_postions, num_mlm_preds, vocab)
+    pred_positions_and_labels = sorted(pred_positions_and_labels, key = lambda x: x[0])
+    pred_positions = [x[0] for x in pred_positions_and_labels]
+    mlm_pred_labels = [x[1] for x in pred_positions_and_labels]
+    return vocab[mlm_input_tokens], pred_positions, vocab[mlm_pred_labels]
+
+# Append special <mask> tokens into the inputs and padding them together
+# exmaples: outputs from  _get_nsp_data_from_paragraph and _get_mlm_data_from_tokens
+def pad_bert_inputs(examples, max_len, vocab):
+    max_num_mlm_preds = round(max_len * 0.15)
+    all_token_ids, all_segments, valid_lens = [], [], []
+    all_pred_positions, all_mlm_weights, all_mlm_labels = [], [], []
+    nsp_labels = []
+    for (token_ids, pred_positions, mlm_pred_label_ids, segments,
+         is_next) in examples:
+        all_token_ids.append(torch.tensor(token_ids + [vocab['<pad>']] * (
+            max_len - len(token_ids)), dtype=torch.long))
+        all_segments.append(torch.tensor(segments + [0] * (
+            max_len - len(segments)), dtype=torch.long))
+        # `valid_lens` excludes count of '<pad>' tokens
+        valid_lens.append(torch.tensor(len(token_ids), dtype=torch.float32))
+        all_pred_positions.append(torch.tensor(pred_positions + [0] * (
+            max_num_mlm_preds - len(pred_positions)), dtype=torch.long))
+        # Predictions of padded tokens will be filtered out in the loss via
+        # multiplication of 0 weights
+        all_mlm_weights.append(
+            torch.tensor([1.0] * len(mlm_pred_label_ids) + [0.0] * (
+                max_num_mlm_preds - len(pred_positions)),
+                dtype=torch.float32))
+        all_mlm_labels.append(torch.tensor(mlm_pred_label_ids + [0] * (
+            max_num_mlm_preds - len(mlm_pred_label_ids)), dtype=torch.long))
+        nsp_labels.append(torch.tensor(is_next, dtype=torch.long))
+    return (all_token_ids, all_segments, valid_lens, all_pred_positions,
+            all_mlm_weights, all_mlm_labels, nsp_labels)
+
+def tokenize(lines, token='word'):
+    # Split text lines into word or character tokens
+    if token == 'word':
+        return [line.split() for line in lines]
+    elif token == 'char':
+        return [list(line) for line in lines]
+    else:
+        print('ERROR: unknown token type: ' + token)
